Improvements to IClickSelectableControl, IsSelectedChanged
I've now added in IsSelectedChanged to the IClickSelectableControl interface, as well as a few other methods. The IsSelectedChanged event is done as a routed event, with a newly-improved ItemSelectionChangedEventArgs class.

This more amounted to a bigger change with FlatButton and SplitButton, but it was done!

From here, I can actually modify SelectPanel to only expect a collection of IClickSelectableControl objects, rather than SelectableUserControls. This will give SelectPanel a lot more flexibility! :D