Various minor performance improvements
I'm going to avoid digging into each individual item to explain the changes. In general, this is a lot of minor performance or edge-case-avoiding improvements, as suggested by the included C# analyzers.

The biggest holdback from me implementing more is that some improvements rely on features added in .NET Core 3.1, which the .NET Framework code won't have access to.