Remove a few .NET Core only suggestions
Removed a few suggestions that can only be applied with .NET Core only code, not .NET Framework code. It's not worth the effort to totally mess up my code's readability for just a little bit of a performance gain.

In the future, I'd more prefer to write my own internal method from where I could easily differentiate between .NET Core and .NET Framework, but that's not a priority of any level.