Apply full TransformEditDialog and TransformEditor
These new features for 2.0 are fairly standalone - pretty much everything else is either heavily tied together or introduces breaking changes. The dialog was already sitting in a half-finished state in this branch, so I figured I may as well just go the full way and bring the completed feature in.

I'll have to do a bit more bug testing just to make sure there isn't anything weird with the dialogs, and then I'll be set here.