Add .NET Framework 4.0 support!
It took a long time and a lot of work, but I finally got it done! Support for .NET Framework 4.0!

The hardest part of this for me was actually handling WindowChrome. In .NET Framework 4.5 and later, WindowChrome is included in the framework under the System.Windows.Shell namespace; in .NET 4.0, that class isn't there. Luckily, an earlier version of it is available via an older library, which I have a copy of, but that meant that I had to create a class to manage handing off the WindowChrome data to either the framework class or the library class based upon what version of the Framework we're using. It was a bit of a headache.

The other major problem is that .NET Framework 4 has much less convenient methods/code for handling Reflection; 4.5 is so much better in this regard. But luckily all the things I still needed were still there, it just took a few extra steps.

Either way, it's all working now!

There are a few regressions I need to still fix:

1. There's no shadow behind the window, and I'm not sure why. I'll have to investigate why that is or what I can do about it... Or maybe it just was that way the whole time lol.

2. The changes I had to make to the ImageBrushEditorDialog surprisingly made it near useless, so I'm going to have to fix that.

Otherwise, pretty much everything else seems to be present and accounted for! Wow!