Remove earlier attempt at getting WindowChrome to work
My original attempt was going to just copy over the WindowChrome file from the modern WPF repo and get it work retroactively with .NET Framework 4. Technically, this would've actually worked just fine, but the amount of time and effort that it would've taken for me to get it all set up was a bit daunting.

Luckily, since I still had that old library sitting around, I didn't have to deal with this step!