Attempts to get drop shadow working on FlatWindow
I'm not super sure why it stopped working. I figure it's something to do with all the messing around I did with WindowChrome to get to this point, but it seems weird that this doesn't have a shadow, and it feels like an annoying regression that others would otherwise not want to have.

I wonder if I just need to do a restart or something. Maybe it's something with my PC.

Anyway, I'm putting this away for now lol.