Created a RectEdit control... for editing Rects
I'm starting the process of updating the ImageBrushEditorDialog, because apparently I just want to do more work for myself lol.

I think this is feature-complete, and you can use LoadRect and GetRect to get Rect values out of it. We'll see in the next commit if this is actually error-free or not.

(Also, prior to the 1.9.7 release, I'll add on a new RectEditor to the PropertyList... because why not lol).