Initial rework of ImageBrushEditorDialog
The RectEdit controls are not rendering like they're supposed to, and I'm not really sure why. I'm going to have to see if this is something I can easily fix; if this takes too much work, I might delay this control until 2.0 so I have more time to workshop it.

Other than that, I moved X alignment and Y alignment up to be always visible, and updated the preview to reflect changes in their values. Turns out, these have an effect on the result, even if you don't mess with the viewbox and viewport.

I'm going to have a fun time picking this apart if I end up not getting RectEdit to work.

Also includes improvements to XML comments in EnumComboBox, but no real changes in functionality.