No RectEdit, more work on ImageBrushEditorDialog
Removed the RectEdit control from the ImageBrushEditorDialog. I don't feel like trying to deal with whatever it has going on right now, I'd rather save that for version 2.0. For now, I'm going to remove RectEdit in a future commit.

I did do some rearranging and improvements to the dialog though! I rearranged how the Viewbox and Viewport controls are done, so that it looks a bit nicer. I also added a Presets menu to quickly set some options.

There's a little bit more I want to do with this dialog, and then I should be set to go!