Add "Set to Full Image Size" to ImageBrushEditorDialog
Added buttons to set the Viewbox or Viewport of an image brush to the full size of the image source. This felt like a nice convenience thing to add, especially before I actually understood what the Viewbox and Viewport did lol.

At this point, I'd consider this dialog as complete as it's gonna be for version 1.9.7; I'll leave myself room to try to get the RectEdit working with it for version 2.0.

This is the last major feature or change though, now it's time for some bug testing before the release! :D

(Also made some adjustments to CharInputDialog's Grid measurements, in case I ever end up allowing the window to be resized.)