More improvements to ImageBrushEditorDialog
I added an icon to the "Select From File" button, to try to make it a bit more noticeable. Don't know if I want to go with this icon, or a more standard "Open" icon...

I also added extra handling for when an image is being downloaded from the Internet. If the download fails, an error message is displayed and the image source is reset to a question mark. If the download succeeds, the dialog refreshes the image source.

Note that .NET Framework 4.0 only supports TLS 1.0, not the newer versions (only 4.8 added TLS 1.3 support), which limits it to pretty much being unable to load in an image from the Internet. So for that specific framework version, I changed the fallback text to not mention entering a URL.