Change CheckBox Check color in High Contrast mode
When in high contrast, I'll make sure that the color of the check doesn't match the color of the background color. This was a problem as the white-on-black and green-on-black themes would have a white checkmark on a white background.

This isn't a perfect solution, as this only checks if the colors are exactly the same, not if the colors are within a low contrast ratio (and if the background isn't even a solid color, then this doesn't have any effect at all). But it's a start that will help with those who just rely upon the default values.