Quickly exit CheckPropertyDisplay if no attributes
If a property has no attributes, then there's no reason to check them. (In the future, maybe another shortcut I can do is check if a property only has 1 attribute that matches a certain common but unrelated type... but that's a thought for another day).

Honestly, anything I can try to do to make it faster to load up objects in the PropertyList would be great! But I think the issues ultimately end up with just having to create so many UI objects in a short period at once.