Fix some code warnings
When I was putting together most of the Ribbon controls and code so far, I haven't actually gone and applied XML comments to most of them. But now I have at least put in some! There's still a looooot of warnings, which I'll probably get to fixing... some other time.

I also enabled nullability for the SsuiSample project. Not quite sure if this was the best route, but whatever lol.