Added LinkTextBlock to Avalonia project
This was my first foray into using the StyledProperty (Avalonia's version of DependencyProperty) and in general interacting with how Avalonia handles these routed properties and events.

In general, I do quite like it! However, it's definitely been a bit of a learning curve, and it's also clear to me that I'm not going to be able to use pretty much any UI control's C# file for both platforms. For the future, I'm going to have to copy over source files and then modify them to make them work with Avalonia.

Thus, this LinkTextBlock is going to sit as the only weird hybrid encapsulating both lol. Perhaps in a future commit I'll split this into a WPF LinkTextBlock and an Avalonia LinkTextBlock.