An attempt to get SSUI working on Windows XP
I tried to copy over SsuiSample over to a Windows XP virtual machine, but SsuiSample didn't actually render correctly - it was mainly just a black window with some UI elements sometimes appearing or disappearing.

This was an attempt that I copied from MessageDialog (because somehow MessageDialog works just fine), but it didn't work for some reason with all FlatWindows.

I tried... For now, I'll have to note that 1.9.7 does not work on Windows XP.

Also, changed the text in the DialogsTest to a LinkTextBlock, so that I can test the MessageDialog.