Add ClickSelectTextBox and AutoCompleteBehavior
These are the last easy controls to port over... everything after this is gonna involve needing setting up control templates and such.

Since the TextChangedEventArgs object is basically useless in Avalonia, the AutoCompleteBehavior class isn't exactly the same as it is in WPF; I'm hoping that it's just less performant, and not just non-functional. Once more stuff gets implemented, I'll get to start testing it.

This whole porting to Avalonia thing has been teaching me a lot!