Add in FlatButton - not quite done, won't build!
Started work on porting over the FlatButton. Still quite a bit more work I have to do, but the "Click / Selection Handling" section is pretty much all set.

Because the code is still missing a lot of needed visual properties and such (and also the XAML bindings aren't set up correctly in the ControlTheme), this won't build as it is right now.

My next commit with this will be alll about finishing up the final needed parts! Once I have this first control done, then I'll have a better idea of what I'll need to do for the rest :)