More work with Avalonia port of FlatButton
Added the needed handler for CanSelect's property changing, and also did a bit of code cleaning up and organizing. At this point, the actual code and functionality is complete, so now it's just about getting the ControlTemplate to work.

Also did a little bit of organizing with the WPF version's code.