Finish up template for NewIntegerSpinner
Now NewIntegerSpinner is all up and working! :D However, pressing and holding a button currently doesn't activate the advance timer, so you're not able to quickly change a value that way. That's something I'm going to have to fix before I can actually go ahead and release it.

Beyond that, though, everything's been pretty much all set up! Including getting all the finicky stuff involving the borders and the corner radius done, pretty much all in XAML! This will help with reducing the amount of C# back-end code actually needed.

As a part of this, I also moved the ColorScheme property to NewSpinnerBase, so that's less code each inheritor needs to set up.

I think what I'll need to do is set up a RepeatFlatButton (so as to not bloat the code of FlatButton itself any further lol).