Finish FlatRepeatButton, NewIntegerSpinner
Changed FlatRepeatButton to use a DispatcherTimer instead, which alleviated the issues I was having. With that, I finally binded the final properties from NewIntegerSpinner to the FlatRepeatButton, and now we're all set! :D

With all this work done with the NewIntegerSpinner, now I can get started with the LongSpinner and DoubleSpinner next!