Change snap points to be dependency properties
Changed HorizontalSnapPoints and VerticalSnapPoints to be dependency properties, in both RelativePositionSelect and TwoPositionsSelect.

Usually collection types are defined as read-only dependency properties, but that would've have allowed them to be set via XAML, which is an aspect that I really wanted, so these are read-write properties (meaning someone can set them to a completely new ObservableCollection). To make sure that I still have access to these collections though, I added a handler for when the properties change to make sure to subscribe and unsubscribe from the CollectionChanged events.