Create StringSpinner control
This is another control that I've now done! This control is able to select a single string from a list of strings, via a spinner/updown control. There is a similar control to this in WinForms, called a DomainUpDown, and I wanted to try to replicate that here.

With the SpinnerBase base class present, that helped with getting a lot of the base work done, but this is still probably not 100% complete. I might try to switch around the arrows/behaviors so that pressing down increases the index rather than decreases. And I'm sure there's some scenarios that this could probably break, but I think it's already relatively robust.