Add AutoCompleteBehavior to StringSpinner
Did a bit more of edge case handling, although there's still probably more work to be done.

Also added on AutoCompleteBehavior to the StringSpinner's TextBox, to make it easier to type in values. Perhaps I should add a way to disable it if desired, but for now, it's just on all the time.