Add CursorEditor, CursorType.ToCursor extension
Add a CursorEditor type to PropertyList. I thought I would've had to postpone this to after 2.0, but this actually wasn't too hard to make happen.

As a part of this, I did also add on an extension method called ToCursor for the CursorType enum - since the actual Cursor constructor that takes the CursorType enum is actually internal to WPF only.

Finally, this also adds on the Select icon, which looks like a mouse cursor.