Added SolidShineUi Styles to make adding even easier!
Now, rather than including a merged resource dictionary in your App.xaml file, you can just easily include a <flat:SolidShineUi /> element in there (and add the "flat" namespace, of course). This is a lot more similar to what you have to do to add on FluentTheme to your application, and means less typing for developers using this package!