Set FlatButton padding in the control's style
This way, this is now set by default for all buttons using this style (which by default, is all buttons), rather than having it set via the constructor. This way, it makes it more predictable for people to edit this property (such as via their own styles).

Later, I'll need to remove this from the FlatButton's constructor.