Begin work on Avalonia version of CreateFromMenu
After investigating how Avalonia does its event handling, I've figured out what I need to do! This is the start of a lower method called CreateFromClickEvent, which will work not just on menus but also buttons or any Avalonia control that has a Click routed event. This method isn't complete yet (as I need to put in the final bits) but once it's done, I'll be able to then build the larger CreateFromMenu method like what's present in the WPF version of this class.