Backport IClickSelectableControl interface from 2.0
Backported one of the new core interfaces from 2.0. This interface will allow me and others to apply more control types to the SelectPanel beyond just a SelectableUserControl, and I feel it was important to end the 1.9.x series of releases with that extra freedom with SelectPanel.

This, however, did require a bit of modification to the code.

Firstly, SelectableUserControl's new IsSelectedChanged event now uses a different (but identical in composition) ItemSelectionChangedEventHandler delegate than the older SelectionChanged event; I only changed IsSelectedChanged as this is the newer event (added in 1.9.7) while I wanted to keep SelectionChanged unchanged for those who upgraded from an older version.

Secondly, I changed FlatButton's IsSelectedChanged event to use the new ItemSelectionChangedEventHandler delegate. Most functionality from the old event handler is still here (just now no access to the Property property, but that can still be accessed via FlatButton.IsSelectedProperty), so I felt okay with making this change, even if it is a breaking change.