Change how FontSelectDialog handles FontWeights
By changing the FontSelectDialog to handle FontWeights by their integer OpenType weight value, rather than directly interacting with the FontWeight type, I've now provided a lot more flexibility for weights that are inbetween the presets listed. I've also been able to avoid a binding issue that the Preview text was having, due to this change.

With this, there should no longer be any errors or crashes while using this dialog.