Fix event handlers in BrushEditor
Because I changed the SelectableUserControl's Click event to be a routed event, this means that I need to update event handlers to use RoutedEventArgs. So this fixes that!

Surprisingly, I think most implementations in my code use SelectionChanged on the parent control, rather than actually listening to a SelectableUserControl's Click event. BrushEditor is unique in that it has a static list of controls.