Move Ribbon to a higher namespace in SolidShineUi
For discoverability purposes, I'm moving the Ribbon to be a namespace immediately under SolidShineUi (so now it's SolidShineUi.Ribbon.Ribbon control, rather than SolidShineUi.Toolbars.Ribbon.Ribbon control).

The Ribbon and its core elements will still be based off of the work for Toolbars (such as IToolbarItem), but for more organic searching through the class library and such, I feel people may be confused as to why Ribbon is considered a subnamespace under Toolbars, when the Ribbon is majorly a completely separate control from a Toolbar control.