More work on dialog launcher for RibbonGroup
Added actual functionality for the launcher now - clicking on the launcher will now activate the LauncherClick event in RibbonGroup. I also finally hooked up the LauncherTooltip and IsLauncherEnabled properties!

In addition, I've also added another launcher button that will be visible on the side of the group if you hide the group header. This is to align with how launchers appear on the single-line Ribbon in Microsoft Office.