Fix compile errors with SsuiSample
I'm actually not 100% sure what happened with the SelectPanelTest. I think it was an issue that arose from me changing git branches, but Visual Studio still holding on to some info from the other branch.

Also, I'll have to port over the CursorEditor change back to the main branch lol.