Add ApplyColorScheme to IPropertyEditor interface
The eventual goal is to remove the ColorScheme property as an interface requirement, since I just don't like the idea of having a set-only property, when just having a method works just as fine.

Of course, since we already have so many property editors, I pretty much had to go through each one and update them.