Organize CharInputDialog, add missing value setters
Organized the code for CharInputDialog and also added/fixed XML comments. I also added on the setters for ValueAsChar and ValueAsRune, as I had seemingly not actually coded those in when I created the dialog.

Hopefully with these additions/changes, this dialog is a bit more useful and friendly for others to use outside of the context of the PropertyList control. (Granted, given the data type that this is built to handle, I'm not sure how often this will come in handy.)