Add resizing/compacting algorithms to Ribbon
Now, if you make the Ribbon's width smaller than what it can do to reasonably fit all of the current tab's commands into it, it'll attempt to compact down groups - like you'd expect a Ribbon to do.

The compacting system that I use has a bit less fine-tuned control over the individual items in a tab than other Ribbon controls may allow (especially compared to how dynamically items can be resized in the Ribbon in the MIcrosoft Office programs), but it does get its job done!

I had some difficulty for a bit since I wasn't getting updated measures after causing a change to a group's compact size mode, but I figured it out.

One thing that is not functioning right now is when a group is compacted down to IconOnly - the icon may appear, but clicking on the icon or text in this state does nothing to make its actual items appear underneath. That'll need to be sorted out in the future.