Add support for RibbonTab.FitContentsToWidth
When FitContentsToWidth is set to false, then the Ribbon will not do compacting for that particular tab. This defeats the need for a Ribbon-wide DoCompacting property (although I still might want to consider that), as this can be controlled on a tab-by-tab basis.

Also expanded the XML comments for various properties in RibbonGroup related to compacting, to try to explain the process.