Change IKeyAction SourceElement to InputElement
For the Avalonia version of the library, I'll actually use InputElement as the type of SourceElement, rather than Control. This way, objects that can still interact with inputs but don't render on the screen can still be used in this way. Also, the naming just makes more sense with this.