Finish RoutedEventKeyAction.CreateFromRoutedEvent
With this, now the convenience method CreateListFromMenu is also now present here in Avalonia. In fact, with CreateFromRoutedEvent, this is a lot more useful and powerful than just CreateListFromMenu.

I think I should make a companion method for CommandKeyAction, for those who are using commands for their UI rather than routed events.