Rename RibbonCustomControl to RibbonContentControl
The rename gives it a more generic sounding name but is also more fitting for the fact that it's able to support any sort of content and put it on the Ribbon. I'll also be using this control as a base for future controls (such as RibbonComboBox), and so a name like "CustomControl" might give a wrong or weird impression.

With this commit, I've also finished up work on the RibbonContentControl's control template. This includes letting the control be full height when the size is set to "Large", and lowering the opacity when the control is set to disabled.