Add RibbonComboBox control
This control displays an actual ComboBox on the Ribbon, by using the RibbonContentControl and setting its content to a ComboBox. All of the properties from the ComboBox are exposed as properties on the RibbonComboBox as well (and using binding in the back-end to keep the two in sync). In the future, I may go through and change this slightly, including adding a custom style to make the ComboBox fit in more with the Ribbon, but for here and now, I'm quite happy!

I'm going to use this same idea to create other  entries later, such as RibbonTextBox and RibbonIntegerSpinner.