Added XML comments to RibbonContentControl, RibbonCheckBox
These are now the only two controls with progress towards filling in the XML comments for their properties; with these as the template, I can use these for filling in the XML comments of the other controls later.

Also fixed a height calculation for RibbonContentControl when it's in large size.