Finish up design of Menu ControlTemplate
I've pretty much now made a good replication of the visuals of the Menu from the WPF version!

An interesting thing to note is that by default, Avalonia seems to like going with a font size 14, in comparison to the font size 12 that WPF (and Windows as a whole) falls back on. So I've updated all of my visuals to use font size 12. I might try to set the default across the library to be font size 12.

In general, Avalonia's menu design seems to be designed to have larger text and larger spacing between items. I'm all for having good spacing, but it was a bit much lol. I think it's because Avalonia is, by default, trying to have a less dense look, which isn't necessarily a bad thing, but isn't what I'm looking for with my UI.

I think what I'll end up doing in the future is making less-dense styles of these controls in the future that are more fitting for font size 14 and the larger spacing that Avalonia's Fluent theme seems to favor, but that's probably beyond the scope of what I want to accomplish for 2.0. Right now, my biggest goal is just getting Solid Shine UI ported over and try to have the visuals match as much as possible.