Add ContextMenu control!
Yet another control in the bag! This was pretty easy to implement by just taking the existing Fluent theme's ContextMenu ControlTheme and updating it to use the FlatMenuScrollViewer, and hooking up all of the properties.

For the ControlTheme, it seems that the border inside the theme is supposed to pull its values from the setters defined just above, but yet they didn't seem to pull in for some reason; so instead, I've just set them to either directly pull from the properties, or just use an existing resource.

In the future, I might want to find ways to improve this, so that more flexibility can be provided. But for now, this is just fine.