Create template dictionary for MenuButton
Created the resoruce dictionary file that I'll use to put in the MenuButton. I don't want to put it into the Buttons.axaml one, so as to avoid circular referencing (the MenuButton will need to refer to the Menus template file, and the Menus template file already refers to the Buttons template file).

I might in the future change "Buttons.axaml" to just "FlatButton.axaml".

Also, the actual control theme isn't present here yet at all. Just wanted to set it up, so that I can add it in a bit later!