Add HorizontalOffset, VerticalOffset for MenuButton
Previously, the HorizontalOffset and VerticalOffset was hardcoded, but now it can be set directly in the menu button, in case you want to override the default values.

I've also expanded the StaysOpen documentation, since there are some differences between the WPF and Avalonia versions of SSUI.

This is being made its own commit so I can cherry-pick it into the main branch later.