Re-added the negative value check in HSV/HSL hue
Re-added the check for negative values used in certain instances when calculating the hue in HSV or HSL. I thought it was unnecessary, but according to the comments in Avalonia's code, there is a reason to avoid having negative values at that specific place. And I guess I determined this years ago, but never wrote a comment to explain that lol.