Initial work on RibbonSplitButton
First commit for the RibbonSplitButton. Actually a large majority of the functionality and work is there, including the actual functionality with the buttons. This code was honestly pretty thrown together, so I'm surprised it doesn't look and function even worse than it does.

That being said, there is still a lot that needs to be done (especially a lot of the layout/spacing issues), and I may have been a bit too overzealous at commenting out or removing things in the goal of just getting this to build. So I'll need to come back through here with a finer attention to detail and a more put-together plan.