Initial work done on SplitButton's control template
Still a lot of work to do (especially in regards to the MenuButtonPlacement property), but at least the basic idea is now present and working pretty well!

That being said, I am having some issues with getting the separator to appear by default, I'm not super sure what's going on. I'll need to mess with this some more later.