Fix separator appearance in SplitButton
In the WPF version, I use a border control to draw the separator, but I was getting inconsistent or weird results here, so instead I'm just using a basic panel. And it seems to be working pretty well! :)

I've set a lot of default values for the brushes/visuals now, which hopefully is okay. Either way, people are definitely able to change these properties to set to whatever they want as needed.