Remove padding from FlatButton constructor
This is a breaking change, but a needed fix on earlier me's silly choice lol.

Previously, Padding was defined in the constructor of FlatButton, meaning that it most likely overwrote the Padding whenever it was defined in XAML anywhere else. Including the styles of my own controls that built upon FlatButton. Not great.

Changing this to instead set the Padding by default in the FlatButton's style solves this problem and allows anyone (including me) to freely modify the Padding at any other time, but it has a nasty knock-on effect affecting every other button control of mine where I inadvertantly relied upon the padding being what it was to have the controls look how they should.

This commit catches and fixes many of those knock-on effects, but there may be more than I was accounting for. And on top of that, this may also have an effect on others' software too when they upgrade to 2.0.

This is a silly mistake that I had made a long time ago, and now I worry it might be a bit messy for anyone who upgrades and was expecting the same behavior.