Fix up ColorScheme/visuals of RibbonSplitButton
I hadn't set up the ColorScheme code quite correctly, which is silly since I've done this so many times before.

Also, I finished setting up the visuals, so the RibbonSplitButton looks great! :) (Except the CornerRadius property, that currently does nothing lol.)

It currently does not do the thing in Office's Ribbon where highlighting the main button will also display the border of the menu button (and vice versa). Right now, I don't see that as a priority, but that should definitely be something that is added on in time.