Created a RibbonTextBox control
Now we can add text boxes to our Ribbon! :D It's the same basic idea as RibbonComboBox, and I had shared on stream that I would be making it on my own (as this wasn't particularly entertaining to do).

Everything that is needed is present though, including a few extra properties that I needed to provide more control over the size of the textbox and how it handles text alignment.