Add CheckBox control to Avalonia
This is just the beginning with this control, but most of the properties I need are here - as well as the logic for when the CheckState is changed and raising the related events and changing the related properties.

I still need to finish the click handling and other things, and then after that actually build the control template. One step at a time though!