Initial work done for CheckBox style
Got the basics put in place for the CheckBox style; there will still be more I need to do (such as supporting the :disabled state), but it is at least more than functional right now and also has some added features not present in the current WPF CheckBox (which I'll need to later on rewrite after I finish this).

As a part of this, I've made some other changes or additions. I've now set a default value for HighlightBrush and BackgroundHighlightBrush, so that the box area in general gets a little bit lighter (increase in lightness) when the pointer is over it - before the only change would be the border around the checkbox.

A new gray color is used for HighlightBrush, which I've stored in ColorsHelper (and I may store other custom colors I develop there as well for future re-usage and to keep it all contained in one place rather than scattering random used colors in various properties or other files).