Minor adjustments to RibbonSplitButton
I'm going to spend some time (not on stream) figuring out how I set up the visuals for this, because I'm not sure if I'm liking the final result that we've got here.

More specifically, this doesn't have the TransparentBack property available at all. While by default, all Ribbon buttons have TransparentBack set to true and are designed with that in mind, having the property available allows developers to highlight a button if they so choose - even if it isn't part of the original guidelines. Sometimes UI is just that way lol.

So I think I'll go back and re-implement a TransparentBack property here, and then update the ControlTemplate so that it accounts for that and sets it to true; this way, it'll also have it set for when a RibbonSplitButton is added to the UI but a ColorScheme is not yet supplied.