Minor progress with RibbonFileMenu
My time here, I think, was more spent on re-learning what I had done thus far with the RibbonFileMenu, less than actually making solid progress with doing something with it. Before I continue working on this more, I'm going to need to spend some time working on how I want to design the file menu and how I want the code for that to look.

In this commit, what I did was add on the ColorScheme property (and have Ribbon fill that in) so that the colors of the menu aren't just a jarring gray - although the buttons in the file menu still aren't updated by the ColorScheme. I've also made some tweaks to the layout and how brushes are applied, which makes for a more solid base for me to work on top of after this.

I've also fixed a bug/bad behavior of the click brush applying to the main button whenever I click within the actual popup menu - now it only applies when I click on the button itself.