Rearranged Ribbon code, added main bar scrolling
Now scroll buttons will appear for the main bar if the content (groups, commands, etc.) is larger than can be shown - even after compacting. There are some situations where the buttons might appear when they really don't need to... I've tried to mitigate the visual and practical issues in that case, but it still happens. But for now, it's good enough (tm).

Also fully rearranged the code in the Ribbon.cs file, so that the regions make a bit more sense. Still kind of a big file with a lot going on though!

Also changed the scroll commands to always be active in the Ribbon and TabControl. Had a weird bug where the buttons were sometimes disabled when they shouldn't have been, and I think it was because WPF hadn't quite yet checked if the commands can be activated yet or not.