Add on the spinner controls for the Ribbon
Now all of the spinner controls are present for the Ribbon! On the previous stream (Ribbon Part 14), I showed me setting up the RibbonSpinnerBase and starting with the RibbonIntegerSpinner. Now I've finished off that control and also added on the other ones.

Right now, the only one I'm really questioning the design of is the RibbonStringSpinner, but to be fair, the base StringSpinner itself I'm also not feeling 100% great about how I designed it either. But that will probably be a thing I address after 2.0.