Begin work on IntegerSpinner
The cool thing seems to be that most of the code that I made for the WPF IntegerSpinner also seems to work with this Avalonia version, so yay! Of course, that's becaues it's primarily built upon the fairly abstracted-out SpinnerBase that I've already made, so the hard stuff is kinda already done lol.

Although the next hard thing to do is actually building the template for this control. That'll come in a near future commit!

Also, Visual Studio has been real laggy while I've been typing this commit message lol