Update to Avalonia 11.1 RC
I wanted to update to 11.1 to try to take advantage of the changes to the menu items that was pulled in (i.e. adding on the check mark and radio mark). As a release candidate, I'm expecting it to be relatively stable, but I'm also expecting that the final 11.1 will be released by the time that this is completed. Once 11.1 is released, that'll be the version that I'm sticking with from here on.

In a future commit, I'll have to look at my control template for MenuItems and compare it against Avalonia 11.1's Fluent Theme template, to see what parts I need to add.