Begin progress on rewriting CheckBox
I rewrote the brushes for the CheckBox control, adding on some of the ones that I added in the Avalonia version. With this, I also decided to no longer have my own overwritten BorderBrush property, and instead just use the base Control property. Also, I removed all remnants of the BorderSelectedBrush, which is not meant to be used.

In the future, the added brushes will be added onto the template, when I get to doing a bit of rearranging there.

Finally finally, I added on two more colors that are stored in ColorsHelper. These are also taken from the Avalonia version.