Remove OnlyAllowCheckBoxClick on CheckBox
After thinking back and forth about it, I've decided that I want to remove this functionality. It adds complexity to the control template (although it really wasn't that complex on the code side), and it also was never a feature that I used at all. This functionality was also unique to this CheckBox, and wasn't present in the WPF version. Finally, I don't see an easily viable way to achieve this with the Avalonia version of this control, and I'm trying to have as few differences as possible.

I believe that the best design for the CheckBox is to allow the label to also be able to trigger the check state change too.

Existing workarounds are present to have child elements fully consume the MouseDown and other events if that is needed, similarly to what would need to be done with the WPF CheckBox right now.