Redesign CheckBox control template
Redesigned the CheckBox control template to make it easier to manage and understand. With the brushes I added on, I was able to replicate the minor highlighting that I had implemented in the Avalonia version.

With the updated brushes and appearance, I also rewrote how the color scheme is applied, which leads to a more consistent and predictable experience there. This is, once again, a change in how the color scheme is calculated for high contrast themes, which I'll revisit again when I redesign how color schemes/themes work.

From here, now that I use a DockPanel, my eventual goal is to add a property to allow the checkbox to be moved to other edges of the control.