Add BoxPlacement property to CheckBox
With this property, you can now move the actual checkbox to the other edges of the control, rather than having it only be stuck to the left. I feel this will be a fairly niche option, but it's at least here for those who want it! Unlike the OnlyAllowCheckBoxClick property, this one is a lot easier to maintain and doesn't require that much extra code at all.

I also fixed a few more things with the control template. I've changed how CheckBox's height is calculated, and set the MinHeight to 22 by default (rather than 24). I also now utilize the VerticalContentAlignment and HorizontalContentAlignment properties where I didn't before. Finally, I added support for access keys in the content. (whoops!)

With all of these options, now other developers should have fairly decent control over how their content is rendered in relation to the checkbox.