Create TimeSpinner control
It is time for a new time-based control! The TimeSpinner allows displaying any values of a TimeSpan, although I'd expect that it'd be primarily used to input values between 00:00 and 23:59 (so that's what I have set as the MinValue and MaxValue by default). This is built to go alongside a DatePicker, to allow full control over selecting a date value and time value for a DateTime.

This is my first pass at all of this, although I think it should be feature-complete. The parsing is done (including allowing parsing from an integer), and it has features to support a custom format string and also support 12-hour clocks (rather than 24-hour by default).